Remove odd methods to load modules for denite
It was intended that the module includes common funcs to use in the
source and the kind. But it has less necessity, so will be removed.
