+++ rplugin/python3/denite/kind/floaterm.py
@@ -1,11 +1,5 @@
-import sys
-from pathlib import Path
-
-sys.path.append(str(Path(__file__).parent.parent.parent.resolve()))
-
 from denite.kind.base import Base
-from denite.util import Nvim, Candidates, UserContext
-from denite_floaterm import Floaterm
+from denite.util import Nvim, UserContext
 
 PREVIEW_FILENAME = "[denite-floaterm-preview]"
 
@@ -17,7 +11,7 @@ class Kind(Base):
         self.name = "floaterm"
         self.default_action = "open"
         self._previewed_bufnr = -1
-        self._floaterm = Floaterm(vim)
+        self._is_nvim = bool(vim.funcs.has("nvim"))
 
     def action_new(self, context: UserContext) -> None:
         self.vim.command("FloatermNew")
@@ -44,22 +38,34 @@ class Kind(Base):
             self.vim.command("pclose!")
             return
 
-        @self._floaterm.restore_window_wrapper
-        def preview() -> None:
-            self.vim.call("denite#helper#preview_file", context, PREVIEW_FILENAME)
-            self.vim.command("wincmd P")
-            self.vim.current.buffer.options["swapfile"] = False
-            self.vim.current.buffer.options["bufhidden"] = "wipe"
-            self.vim.current.buffer.options["buftype"] = "nofile"
+        self._save_win()
 
-            buf = self.vim.buffers[bufnr]
-            last_line = len(buf) - 1
-            last_non_empty_line = next(
-                filter(lambda x: buf[x] != "", range(last_line, 0, -1)), last_line
-            )
-            start = max(0, last_non_empty_line - self.vim.options["previewheight"] + 1)
-            end = last_non_empty_line + 1
-            self.vim.current.buffer[:] = buf[start:end]
+        self.vim.call("denite#helper#preview_file", context, PREVIEW_FILENAME)
+        self.vim.command("wincmd P")
+        self.vim.current.buffer.options["swapfile"] = False
+        self.vim.current.buffer.options["bufhidden"] = "wipe"
+        self.vim.current.buffer.options["buftype"] = "nofile"
 
-        preview()
+        buf = self.vim.buffers[bufnr]
+        last_line = len(buf) - 1
+        last_non_empty_line = next(
+            filter(lambda x: buf[x] != "", range(last_line, 0, -1)), last_line
+        )
+        start = max(0, last_non_empty_line - self.vim.options["previewheight"] + 1)
+        end = last_non_empty_line + 1
+        self.vim.current.buffer[:] = buf[start:end]
+
+        self._restore_win()
         self._previewed_bufnr = bufnr
+
+    def _save_win(self) -> None:
+        if self._is_nvim:
+            self._current_window = self.vim.current.window
+        else:
+            self._current_window = self.vim.funcs.win_getid()
+
+    def _restore_win(self) -> None:
+        if self._is_nvim:
+            self.vim.current.window = self._current_window
+        else:
+            self.vim.funcs.win_gotoid(self._current_window)
